Fix flaky profiler and test_callback_simple RPC tests
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41287

Profiler tests that test profiling with builtin functions and `test_callback_simple` test has been broken for a while. This diff fixes that by preferring c10 ops to non-c10 ops in our operation matching logic.

The result of this is that these ops go through the c10 dispatch and thus have profiling enabled. For `test_callback_simple` this results in the effect that we choose `aten::add.Tensor` over `aten::add.Int` which fixes the type issue.

Differential Revision: [D22489197](https://our.internmc.facebook.com/intern/diff/D22489197/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D22489197/)!
ghstack-source-id: 107697506